Move Colonini parser-only files to shorter path
