+++ src/bh/config/colonini/internal/parser.h
@@ -25,7 +25,7 @@
 #include <stddef.h>
 
 #include "bh/config/colonini/internal/lexer.h"
-#include "bh/config/colonini/internal/parser_type/assign_statement.h"
+#include "bh/config/colonini/internal/parser/assign_statement.h"
 
 #ifdef __cplusplus
 extern "C" {
