+++ src/bh/config/colonini/internal/parser/const_expr.c
@@ -19,7 +19,7 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#include "bh/config/colonini/internal/parser_type/const_expr.h"
+#include "bh/config/colonini/internal/parser/const_expr.h"
 
 #include <assert.h>
 #include <stddef.h>
@@ -27,7 +27,7 @@
 #include <string.h>
 
 #include "bh/config/colonini/internal/lexer/lexer_string.h"
-#include "bh/config/colonini/internal/parser_type/const_expr_type.h"
+#include "bh/config/colonini/internal/parser/const_expr_type.h"
 
 /**
  * External
