+++ src/bh/config/colonini/internal/parser/key_expr.c
@@ -19,15 +19,15 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#include "bh/config/colonini/internal/parser_type/key_expr.h"
+#include "bh/config/colonini/internal/parser/key_expr.h"
 
 #include <assert.h>
 #include <stddef.h>
 #include <stdlib.h>
 
 #include "bh/common/string_util/memstring.h"
-#include "bh/config/colonini/internal/parser_type/const_expr.h"
-#include "bh/config/colonini/internal/parser_type/subscript.h"
+#include "bh/config/colonini/internal/parser/const_expr.h"
+#include "bh/config/colonini/internal/parser/subscript.h"
 
 static int Peek(
     const struct LexerString* begin_src,
