+++ src/bh/config/colonini/internal/parser/key_expr.h
@@ -1,4 +1,3 @@
-
 /**
  * SlashDiablo Maphack
  * Copyright (C) 2012-2022  SlashDiablo Team
@@ -20,33 +19,43 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#ifndef BH_CONFIG_COLONINI_INTERNAL_PARSER_TYPE_CONST_EXPR_TYPE_H_
-#define BH_CONFIG_COLONINI_INTERNAL_PARSER_TYPE_CONST_EXPR_TYPE_H_
+#ifndef BH_CONFIG_COLONINI_INTERNAL_PARSER_KEY_EXPR_H_
+#define BH_CONFIG_COLONINI_INTERNAL_PARSER_KEY_EXPR_H_
 
 #include <stddef.h>
+
+#include "bh/config/colonini/internal/lexer/lexer_string.h"
+#include "bh/config/colonini/internal/parser/const_expr.h"
+#include "bh/config/colonini/internal/parser/subscript.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif  /* __cplusplus */
 
-enum ConstExprType {
-  ConstExprType_kUnspecified,
-  ConstExprType_kInvalid,
-
-  ConstExprType_kString,
-  ConstExprType_kSignedInt,
-  ConstExprType_kUnsignedInt,
-  ConstExprType_kBoolean
+struct KeyExpr {
+  struct ConstExpr constexpr;
+  struct Subscript* subscripts;
+  size_t subscripts_count;
 };
 
-const char* ConstExprType_GetDisplayName(
-    enum ConstExprType type, size_t* length);
+/**
+ * Starting from begin_src, parses a KeyExpr constituted of tokens and
+ * whitespace no further than end_src. Returns a non-NULL on success,
+ * or else returns NULL and sets error column.
+ */
+struct KeyExpr* KeyExpr_Parse(
+    struct KeyExpr* expr,
+    const struct LexerString* begin_src,
+    const struct LexerString* end_src,
+    size_t* error_column);
 
-enum ConstExprType ConstExprType_MatchString(
-    const char* str, size_t length);
+/**
+ * Deinitializes a KeyExpr, freeing up resources that were allocated.
+ */
+void KeyExpr_Deinit(struct KeyExpr* expr);
 
 #ifdef __cplusplus
 }  /* extern "C" */
 #endif  /* __cplusplus */
 
-#endif  /* BH_CONFIG_COLONINI_INTERNAL_PARSER_TYPE_CONST_EXPR_TYPE_H_ */
+#endif  /* BH_CONFIG_COLONINI_INTERNAL_PARSER_KEY_EXPR_H_ */
