+++ src/bh/config/colonini/internal/parser/subscript.c
@@ -19,14 +19,14 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#include "bh/config/colonini/internal/parser_type/subscript.h"
+#include "bh/config/colonini/internal/parser/subscript.h"
 
 #include <assert.h>
 #include <stddef.h>
 #include <string.h>
 
 #include "bh/config/colonini/internal/lexer/lexer_string.h"
-#include "bh/config/colonini/internal/parser_type/const_expr.h"
+#include "bh/config/colonini/internal/parser/const_expr.h"
 
 static const char kLBracket[] = "[";
 static const char kRBracket[] = "]";
