+++ src/bh/config/colonini/internal/parser/toggle_expr.c
@@ -19,13 +19,13 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#include "bh/config/colonini/internal/parser_type/toggle_expr.h"
+#include "bh/config/colonini/internal/parser/toggle_expr.h"
 
 #include <stddef.h>
 #include <string.h>
 
 #include "bh/config/colonini/internal/lexer/lexer_string.h"
-#include "bh/config/colonini/internal/parser_type/const_expr.h"
+#include "bh/config/colonini/internal/parser/const_expr.h"
 
 static const char kComma[] = ",";
